Moved more tests around
