+++ sql/expression/function/functionregistry_test.go
@@ -1,4 +1,4 @@
-// Copyright 2020-2021 Dolthub, Inc.
+// Copyright 2021 Dolthub, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package sql_test
+package function_test
 
 import (
 	"testing"
 
-	"github.com/dolthub/go-mysql-server/test"
+	"github.com/dolthub/go-mysql-server/sql/expression/function"
 	"github.com/stretchr/testify/require"
 
 	"github.com/dolthub/go-mysql-server/sql"
@@ -27,15 +27,15 @@ import (
 func TestFunctionRegistry(t *testing.T) {
 	require := require.New(t)
 
-	c := test.NewCatalog(sql.NewDatabaseProvider())
+	reg := function.NewRegistry()
 	name := "func"
 	var expected sql.Expression = expression.NewStar()
-	c.RegisterFunction(sql.Function1{
+	reg.Register(sql.Function1{
 		Name: name,
 		Fn:    func(ctx *sql.Context, arg sql.Expression) sql.Expression { return expected },
 	})
 
-	f, err := c.Function(name)
+	f, err := reg.Function(name)
 	require.NoError(err)
 
 	ctx := sql.NewEmptyContext()
@@ -55,8 +55,8 @@ func TestFunctionRegistry(t *testing.T) {
 func TestFunctionRegistryMissingFunction(t *testing.T) {
 	require := require.New(t)
 
-	c := test.NewCatalog(sql.NewDatabaseProvider())
-	f, err := c.Function("func")
+	reg := function.NewRegistry()
+	f, err := reg.Function("func")
 	require.Error(err)
 	require.Nil(f)
 }
